+++ clients/algoliasearch-client-csharp/algoliasearch/Utils/QueryStringHelper.cs
@@ -35,18 +35,18 @@ internal static class QueryStringHelper
       case bool boolean:
         return boolean ? "true" : "false";
       case ICollection collection:
-      {
-        var entries = new List<string>();
-        foreach (var entry in collection)
-          entries.Add(ParameterToString(entry));
-        return string.Join(",", entries);
-      }
+        {
+          var entries = new List<string>();
+          foreach (var entry in collection)
+            entries.Add(ParameterToString(entry));
+          return string.Join(",", entries);
+        }
       case Enum when HasEnumMemberAttrValue(obj):
         return GetEnumMemberAttrValue(obj);
       case AbstractSchema schema when obj.GetType().IsClass:
-      {
-        return ParameterToString(schema.ActualInstance);
-      }
+        {
+          return ParameterToString(schema.ActualInstance);
+        }
       default:
         return Convert.ToString(obj, CultureInfo.InvariantCulture);
     }
@@ -63,7 +63,8 @@ internal static class QueryStringHelper
       throw new ArgumentNullException(nameof(enumVal));
     var enumType = enumVal.GetType();
     var memInfo = enumType.GetMember(enumVal.ToString() ?? throw new InvalidOperationException());
-    var attr = memInfo.FirstOrDefault()?.GetCustomAttributes(false).OfType<JsonPropertyNameAttribute>().FirstOrDefault();
+    var attr = memInfo.FirstOrDefault()?.GetCustomAttributes(false).OfType<JsonPropertyNameAttribute>()
+      .FirstOrDefault();
     return attr != null;
   }
 
@@ -78,7 +79,8 @@ internal static class QueryStringHelper
       throw new ArgumentNullException(nameof(enumVal));
     var enumType = enumVal.GetType();
     var memInfo = enumType.GetMember(enumVal.ToString() ?? throw new InvalidOperationException());
-    var attr = memInfo.FirstOrDefault()?.GetCustomAttributes(false).OfType<JsonPropertyNameAttribute>().FirstOrDefault();
+    var attr = memInfo.FirstOrDefault()?.GetCustomAttributes(false).OfType<JsonPropertyNameAttribute>()
+      .FirstOrDefault();
     return attr?.Name;
   }
 }
