+++ templates/csharp/tests/client/suite.mustache
@@ -3,6 +3,7 @@ using Algolia.Search.Models.{{clientPrefix}};
 using Algolia.Search.Models.Common;
 using Algolia.Search.Http;
 using Algolia.Search.Transport;
+using Algolia.Search.Tests.Utils;
 using Algolia.Search.Serializer;
 using System.Text.RegularExpressions;
 using System.Text.Json;
