+++ templates/csharp/tests/requests/requests.mustache
@@ -2,6 +2,7 @@ using Algolia.Search.Http;
 using Algolia.Search.Clients;
 using Algolia.Search.Models.{{clientPrefix}};
 using Algolia.Search.Serializer;
+using Algolia.Search.Tests.Utils;
 using Xunit;
 using System.Text.Json;
 using Quibble.Xunit;
